A 'runslog' command to easily run a slog script locally or on a server
