+++ .pylintrc
@@ -3,7 +3,7 @@
 # A comma-separated list of package or module names from where C extensions may
 # be loaded. Extensions are loading into the active Python interpreter and may
 # run arbitrary code
-extension-pkg-whitelist=protobufs,protobufs.slog_pb2,protobufs.slog_pb2_grpc
+extension-pkg-whitelist=slog.protobufs,slog.protobufs.slog_pb2,slog.protobufs.slog_pb2_grpc
 
 [TYPECHECK]
 
@@ -11,9 +11,9 @@ extension-pkg-whitelist=protobufs,protobufs.slog_pb2,protobufs.slog_pb2_grpc
 # (useful for modules/projects where namespaces are manipulated during runtime
 # and thus existing member attributes cannot be deduced by static analysis. It
 # supports qualified module names, as well as Unix pattern matching.
-ignored-modules=protobufs,protobufs.slog_pb2,protobufs.slog_pb2_grpc
+ignored-modules=slog.protobufs,slog.protobufs.slog_pb2,slog.protobufs.slog_pb2_grpc
 
 # List of classes names for which member attributes should not be checked
 # (useful for classes with attributes dynamically set). This supports can work
 # with qualified names.
-ignored-classes=protobufs,protobufs.slog_pb2,protobufs.slog_pb2_grpc
+ignored-classes=slog.protobufs,slog.protobufs.slog_pb2,slog.protobufs.slog_pb2_grpc
