+++ Dockerfile
@@ -6,9 +6,9 @@ RUN wget -O - https://apt.llvm.org/llvm-snapshot.gpg.key| apt-key add -
 RUN add-apt-repository ppa:plt/racket
 RUN apt-get update -y
 RUN apt-get install -y clang-format clang-tidy clang-tools clang clangd libc++-dev libc++1 libc++abi-dev \
-            libc++abi1 libclang-dev libclang1 liblldb-dev libllvm-ocaml-dev libomp-dev libomp5 lld lldb \ 
-            llvm-dev llvm-runtime llvm python-clang mcpp cmake racket build-essential mpich
-
+            libc++abi1 libclang-dev libclang1 liblldb-dev libomp-dev libomp5 lld lldb \
+            llvm-dev llvm-runtime llvm python-clang mcpp cmake racket build-essential mpich z3 \
+            git python3-pip sqlite3
 RUN raco setup --doc-index --force-user-docs
 RUN raco pkg install --batch --deps search-auto binaryio graph
 
@@ -28,18 +28,16 @@ RUN apt-get install -y python3-pip sqlite3
 
 ENV OMPI_ALLOW_RUN_AS_ROOT=1
 ENV OMPI_ALLOW_RUN_AS_ROOT_CONFIRM=1
+ENV CC=mpicc
+ENV CXX=mpicxx
 
 COPY . /slog
 
 # build backend
-RUN rm -rf /slog/backend/build
-RUN cd /slog/backend && cmake -Bbuild .
-RUN cd /slog/backend/build && make -j8
+RUN cd /slog/backend ; rm -rf build ; cmake -Bbuild . && cd build ; make -j8;
 
 WORKDIR /slog
 RUN pip3 install -r requirements.txt
 EXPOSE 5108
 
-ENV CC=mpicc
-ENV CXX=mpicxx
-# ENTRYPOINT [ "./slog-server" ]
+ENTRYPOINT ["/slog/slog-server"]
