+++ compiler/src/cmake-template
@@ -1,20 +1,9 @@
-cmake_minimum_required (VERSION 2.6)
+cmake_minimum_required (VERSION 3.9)
+
 project (~a)
-add_compile_options(--std=c++17 -Wno-strict-aliasing -fpermissive)
-if(APPLE)
-    set(CMAKE_C_COMPILER clang)
-    set(CMAKE_CXX_COMPILER clang++)
-    if(CMAKE_C_COMPILER_ID MATCHES "Clang\$")
-        set(OpenMP_C_FLAGS "-Xpreprocessor -fopenmp")
-        set(OpenMP_C_LIB_NAMES "omp")
-        set(OpenMP_omp_LIBRARY omp)
-    endif()
-    if(CMAKE_CXX_COMPILER_ID MATCHES "Clang\$")
-        set(OpenMP_CXX_FLAGS "-Xpreprocessor -fopenmp")
-        set(OpenMP_CXX_LIB_NAMES "omp")
-        set(OpenMP_omp_LIBRARY omp)
-    endif()
-endif()
+
+add_compile_options(--std=c++17 -Wno-strict-aliasing -Werror=class-memaccess -fpermissive)
+
 find_package(MPI REQUIRED)
 find_package(OpenMP)
 if (OPENMP_FOUND)
@@ -22,12 +11,16 @@ if (OPENMP_FOUND)
     set (CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} ${OpenMP_CXX_FLAGS}")
     set (CMAKE_EXE_LINKER_FLAGS "${CMAKE_EXE_LINKER_FLAGS} ${OpenMP_EXE_LINKER_FLAGS}")
 endif()
-set (CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -Wall -std=c++17")
-set (source_dir "${PROJECT_SOURCE_DIR}")
-set (ra_source_dir "../../backend/src")
-file (GLOB source_files "*.cpp")
-add_executable(~a ${source_files})
+
+set (CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -Wall -std=c++17 -Werror=class-memaccess -fpermissive")
+set (base_dir "${PROJECT_SOURCE_DIR}/../backend")
+set (source_dir "${base_dir}/src")
+
+file (GLOB source_files_parallel_RA "${source_dir}/parallel_RA_inc.h" "${source_dir}/log/logger.cpp" "${source_dir}/hash/hash.h" "${source_dir}/buffer/vector_buffer.cpp" "${source_dir}/comm/comm.cpp" "${source_dir}/relation/google_btree_relation.cpp" "${source_dir}/relation/shmap_relation_exp.cpp" "${source_dir}/comm/all_to_all_comm.cpp" "${source_dir}/comm/all_to_allv_comm.cpp" "${source_dir}/IO/parallel_io.cpp" "${source_dir}/RA/parallel_join.cpp" "${source_dir}/comm/intra_bucket_comm.cpp" "${source_dir}/RA/parallel_copy.cpp" "${source_dir}/RA/parallel_copy_filter.cpp" "${source_dir}/RA/parallel_copy_generate.cpp" "${source_dir}/RA/parallel_RA.h" "${source_dir}/RA/parallel_acopy.cpp" "${source_dir}/relation/balanced_hash_relation.cpp" "${source_dir}/relation/relation_load_balancer.cpp" "${source_dir}/RAM/RA_tasks.cpp" "${source_dir}/lie/lie.cpp")
+file (GLOB source_files_~a "${PROJECT_SOURCE_DIR}/~a.cpp")
+
+ADD_LIBRARY(parallel_RA "${source_files_parallel_RA}")
+
+add_executable(~a ${source_files_~a})
 INCLUDE_DIRECTORIES(${MPI_INCLUDE_PATH})
-TARGET_INCLUDE_DIRECTORIES(~a PUBLIC "../../backend/src")
-TARGET_LINK_DIRECTORIES(~a PUBLIC "../../backend/build")
-TARGET_LINK_LIBRARIES(~a ${MPI_LIBRARIES} parallel_RA)
+TARGET_LINK_LIBRARIES(~a parallel_RA ${MPI_LIBRARIES})
