+++ requirements.txt
@@ -6,3 +6,5 @@ prompt_toolkit
 yaspin
 pyfiglet
 grpc_interceptor
+six
+sexpdata
