+++ slog-repl
@@ -1,1 +1,6 @@
-python3 -m repl.repl
+#!/usr/bin/env bash
+
+# takes server location as an optional argument
+# defaults to 'localhost' if not provided.
+
+python3 -m slog.repl.repl $1
