+++ slog/daemon/README.md
