+++ slog/daemon/server.py
@@ -13,10 +13,10 @@ import traceback
 import grpc
 from grpc_interceptor import ServerInterceptor
 
-from daemon.const import PORT
-from daemon.rpc import CommandService
-from daemon.task import CompileTask, RunTask
-import protobufs.slog_pb2_grpc as slog_pb2_grpc
+from slog.daemon.const import PORT
+from slog.daemon.rpc import CommandService
+from slog.daemon.task import CompileTask, RunTask
+import slog.protobufs.slog_pb2_grpc as slog_pb2_grpc
 
 class ExceptionToStatusInterceptor(ServerInterceptor):
     """
@@ -24,11 +24,11 @@ class ExceptionToStatusInterceptor(ServerInterceptor):
     grpc server will eat all error silently make server too hard to debug
     """
     def intercept(
-        self,
-        method,
-        request,
-        context: grpc.ServicerContext,
-        method_name: str,
+            self,
+            method,
+            request,
+            context: grpc.ServicerContext,
+            _method_name: str,
     ):
         try:
             return method(request, context)
