+++ slog/daemon/util.py
