+++ slog/repl/completer.py
